OPS-8 frontend heroku deployment terraform

+++ backend/scripts/seed/seed_nibbles.py
@@ -43,29 +43,30 @@ region = os.environ["AWS_REGION"]
 image_bucket = f"{account_id}-{environment_namespace}-nibble-images".replace("_", "-")
 
 
-def upload_nibble_image(image_key):
+def upload_nibble_image(image_key, check_existence=True):
     file_name = f"imgs/nibbles/{image_key}"
     key = f"seeding/{image_key}"
 
     s3_client = boto3.client("s3")
-    try:
-        s3_client.head_object(Bucket=image_bucket, Key=key)
-        print(f"Checked  {file_name}")
-        logging.info(f"{file_name} exists on S3")
-    except ClientError:
+    if check_existence:
         try:
-            s3_client.upload_file(
-                file_name,
-                image_bucket,
-                f"seeding/{image_key}",
-                ExtraArgs={"ACL": "public-read"},
-            )
-            print(f"Uploaded {file_name}")
-            logging.info(f"Uploaded {file_name} to S3")
-        except Exception:
-            print("Error uploading")
-            logging.error(f"Error uploading {file_name}")
-            raise RuntimeError()
+            s3_client.head_object(Bucket=image_bucket, Key=key)
+            print(f"Checked  {file_name}")
+            logging.info(f"{file_name} exists on S3")
+        except ClientError:
+            try:
+                s3_client.upload_file(
+                    file_name,
+                    image_bucket,
+                    f"seeding/{image_key}",
+                    ExtraArgs={"ACL": "public-read"},
+                )
+                print(f"Uploaded {file_name}")
+                logging.info(f"Uploaded {file_name} to S3")
+            except Exception:
+                print("Error uploading")
+                logging.error(f"Error uploading {file_name}")
+                raise RuntimeError()
 
     return {
         "bucket": image_bucket,
@@ -84,13 +85,19 @@ def main():
     with open("data/nibbles.json", "r") as f:
         partial_nibbles = json.load(f)
 
+    # cache images already uploaded
+    updated_images = set()
+
     # fill out remaining nibble info
-    nibbles = []
-    for n in partial_nibbles:
-        if n["imageName"] == "PLACEHOLDER":
+    def full_nibble(partial_nibble):
+        image_name = partial_nibble["imageName"]
+        if image_name == "PLACEHOLDER":
             imageUrl = {"bucket": "PLACEHOLDER", "region": "", "key": "hero"}
         else:
-            imageUrl = upload_nibble_image(n["imageName"])
+            imageUrl = upload_nibble_image(
+                image_name, check_existence=image_name not in updated_images
+            )
+            updated_images.add(image_name)
 
         now = datetime.datetime.now()
         now += datetime.timedelta(
@@ -98,23 +105,21 @@ def main():
         )
         then = now + datetime.timedelta(hours=8)
 
-        nibbles.append(
-            {
-                "name": n["name"],
-                "type": n["type"],
-                "count": n["count"] + random.randint(-1, 5),
-                "imageUrl": imageUrl,
-                "description": n["description"],
-                "price": n["price"] + 5 * random.randint(-10, 10),
-                "availableFrom": int(now.timestamp()),
-                "availableTo": int(then.timestamp()),
-            }
-        )
+        return {
+            "name": partial_nibble["name"],
+            "type": partial_nibble["type"],
+            "count": partial_nibble["count"] + random.randint(-1, 5),
+            "imageUrl": imageUrl,
+            "description": partial_nibble["description"],
+            "price": partial_nibble["price"] + 5 * random.randint(-10, 10),
+            "availableFrom": int(now.timestamp()),
+            "availableTo": int(then.timestamp()),
+        }
 
     # send nibbles
     client = boto3.client("lambda", region_name=region)
     for username in usernames:
-        for payload in random.sample(nibbles, 2):
+        for partial_nibble in random.sample(partial_nibbles, 2):
             response = client.invoke(
                 FunctionName=f"{environment_namespace}-resolver_admin_nibble_mutation",
                 InvocationType="RequestResponse",
@@ -122,7 +127,7 @@ def main():
                     {
                         "field": "adminCreateNibble",
                         "identity": {"username": username},
-                        "arguments": {"input": payload},
+                        "arguments": {"input": full_nibble(partial_nibble)},
                     }
                 ),
                 LogType="Tail",
